+++ src/main/resources/db/migration/20181017150034578__add_template_type_column.sql
@@ -1,0 +1,9 @@
+ALTER TABLE transfer_properties
+    ADD transferType VARCHAR(50);
+
+UPDATE transfer_properties
+    SET transferType = 'ORDER';
+
+ALTER TABLE transfer_properties
+  ALTER COLUMN transferType SET NOT NULL;
+